Int y Numeral Separados
asdasdas

+++ src/parser/Tokens.ts
@@ -33,7 +33,7 @@ export const tokens = {
   // Atoms
   float:        { match: /[0-9]+\.[0-9]+(?:[eE][-+]?[0-9]+)?/, value: (x: string) => (parseFloat(x)) },
   hex:          { match: /0[xX][0-9a-f-A-F]+/, value: (x: string) => (parseInt(x,16)) },
-  integer:      { match: /[0-9]+/, value: (x: string) => (parseFloat(x)) },
+  integer:      { match: /[0-9]+/, value: (x: string) => (parseInt(x)) },
 
   // Identifiers
   identifier:   /[a-zA-Z_][a-zA-Z0-9_]*/,
